Initialize va_list using memset to avoid "possibly used uninitialized" warning Cast to avoid warning on size_t arguments in printf

+++ trunk/cgreen/src/message_formatting.c
@@ -59,9 +59,9 @@ void format_validation_failure_message_for(char *message, size_t message_size, c
 
         if (constraint->size_of_expected_value <= 0) {
             snprintf(message + strlen(message), message_size - strlen(message) - 1,
-                    "\t\tWanted to compare contents with [%s], but [%d] was given for the comparison size.\n",
+                    "\t\tWanted to compare contents with [%s], but [%lu] was given for the comparison size.\n",
                         compared_to_name,
-                        constraint->size_of_expected_value);
+                     (unsigned long)constraint->size_of_expected_value);
             return;
         }
